requirements.txt
ecdsa
pysha3
requests